feat(signup): add signup page for /signup route
