+++ packages/frontend/src/components/signup/forms/TeacherSignUpForm/index.ts
@@ -1,0 +1,3 @@
+import TeacherSignUpForm from './TeacherSignUpForm';
+
+export default TeacherSignUpForm;
