+++ packages/frontend/src/components/signup/forms/index.ts
@@ -1,0 +1,2 @@
+/* eslint-disable import/prefer-default-export */
+export { default as TeacherSignUpForm } from './TeacherSignUpForm';
