--- packages/frontend/src/components/teacher/forms/sign-up/index.ts
@@ -1,3 +1,0 @@
-import TeacherSignUp from './sign-up';
-
-export default TeacherSignUp;
